requirements.txt
streamlit
requests
pandas
google-generativeai
pytz